[sim] Fix probe score events handling
Probe score events were not correctly handed. They were not received
correctly, and when received were handled too late to be displayed.

Correct score probe step.

Introduce a new simulation event to separate the completion and the
finalization of a session. A session is completed when the process is
finished. A session is finalized when it is completed and a last probe
score event has been finalized (if the session is associated with a
probe).